Move test programs source code into testexe/
