+++ src/Pengputer/testexe/HelloWorld.ts
@@ -1,5 +1,5 @@
 import { type Executable } from "@FileSystem/fileTypes";
-import { type PC } from "./PC";
+import { type PC } from "../PC";
 
 export class HelloWorld implements Executable {
     private pc: PC;
